+++ rust/operator-binary/src/resource/listener.rs
@@ -39,7 +39,7 @@ pub fn build_broker_rolegroup_bootstrap_listener(
             .name(kafka.bootstrap_service_name(rolegroup))
             .ownerreference_from_resource(kafka, None, Some(true))
             .context(ObjectMissingMetadataForOwnerRefSnafu)?
-            .with_recommended_labels(build_recommended_labels(
+            .with_recommended_labels(&build_recommended_labels(
                 kafka,
                 KAFKA_CONTROLLER_NAME,
                 &resolved_product_image.app_version_label_value,