+++ rust/operator-binary/src/resource/service.rs
@@ -46,7 +46,7 @@ pub fn build_rolegroup_headless_service(
             .name(rolegroup.rolegroup_headless_service_name())
             .ownerreference_from_resource(kafka, None, Some(true))
             .context(ObjectMissingMetadataForOwnerRefSnafu)?
-            .with_recommended_labels(build_recommended_labels(
+            .with_recommended_labels(&build_recommended_labels(
                 kafka,
                 KAFKA_CONTROLLER_NAME,
                 &resolved_product_image.app_version_label_value,
@@ -87,7 +87,7 @@ pub fn build_rolegroup_metrics_service(
             .name(rolegroup.rolegroup_metrics_service_name())
             .ownerreference_from_resource(kafka, None, Some(true))
             .context(ObjectMissingMetadataForOwnerRefSnafu)?
-            .with_recommended_labels(build_recommended_labels(
+            .with_recommended_labels(&build_recommended_labels(
                 kafka,
                 KAFKA_CONTROLLER_NAME,
                 &resolved_product_image.app_version_label_value,